Move get-version and get-edition implementation to local tool
