+++ cli/src/main/java/com/devonfw/tools/ide/tool/ToolCommandlet.java
@@ -13,7 +13,6 @@ import com.devonfw.tools.ide.process.ProcessMode;
 import com.devonfw.tools.ide.property.StringProperty;
 import com.devonfw.tools.ide.version.VersionIdentifier;
 
-import java.io.IOException;
 import java.nio.file.Files;
 import java.nio.file.Path;
 import java.util.List;
@@ -218,70 +217,12 @@ public abstract class ToolCommandlet extends Commandlet implements Tags {
   /**
    * @return the currently installed {@link VersionIdentifier version} of this tool or {@code null} if not installed.
    */
-  public VersionIdentifier getInstalledVersion() {
-
-    return getInstalledVersion(this.context.getSoftwarePath().resolve(getName()));
-  }
-
-  /**
-   * @param toolPath the installation {@link Path} where to find the version file.
-   * @return the currently installed {@link VersionIdentifier version} of this tool or {@code null} if not installed.
-   */
-  protected VersionIdentifier getInstalledVersion(Path toolPath) {
-
-    if (!Files.isDirectory(toolPath)) {
-      this.context.debug("Tool {} not installed in {}", getName(), toolPath);
-      return null;
-    }
-    Path toolVersionFile = toolPath.resolve(IdeContext.FILE_SOFTWARE_VERSION);
-    if (!Files.exists(toolVersionFile)) {
-      Path legacyToolVersionFile = toolPath.resolve(IdeContext.FILE_LEGACY_SOFTWARE_VERSION);
-      if (Files.exists(legacyToolVersionFile)) {
-        toolVersionFile = legacyToolVersionFile;
-      } else {
-        this.context.warning("Tool {} is missing version file in {}", getName(), toolVersionFile);
-        return null;
-      }
-    }
-    try {
-      String version = Files.readString(toolVersionFile).trim();
-      return VersionIdentifier.of(version);
-    } catch (IOException e) {
-      throw new IllegalStateException("Failed to read file " + toolVersionFile, e);
-    }
-  }
+  public abstract VersionIdentifier getInstalledVersion();
 
   /**
    * @return the installed edition of this tool or {@code null} if not installed.
    */
-  public String getInstalledEdition() {
-
-    return getInstalledEdition(this.context.getSoftwarePath().resolve(getName()));
-  }
-
-  /**
-   * @param toolPath the installation {@link Path} where to find currently installed tool. The name of the parent directory of the real path corresponding to
-   * the passed {@link Path path} must be the name of the edition.
-   * @return the installed edition of this tool or {@code null} if not installed.
-   */
-  public String getInstalledEdition(Path toolPath) {
-
-    if (!Files.isDirectory(toolPath)) {
-      this.context.debug("Tool {} not installed in {}", getName(), toolPath);
-      return null;
-    }
-    try {
-      String edition = toolPath.toRealPath().getParent().getFileName().toString();
-      if (!this.context.getUrls().getSortedEditions(getName()).contains(edition)) {
-        edition = getEdition();
-      }
-      return edition;
-    } catch (IOException e) {
-      throw new IllegalStateException("Couldn't determine the edition of " + getName() + " from the directory structure of its software path " + toolPath
-          + ", assuming the name of the parent directory of the real path of the software path to be the edition " + "of the tool.", e);
-    }
-
-  }
+  public abstract String getInstalledEdition();
 
   /**
    * Uninstalls the {@link #getName() tool}.
